Added timezone.go and moved production code
